requirements.txt
dominate==2.1.12
Flask==0.10.1
Flask-Bootstrap==3.3.5.3
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
python-firebase==1.2
requests==2.7.0
Werkzeug==0.10.4
wheel==0.24.0
WTForms==2.0.2
